+++ src/nodes/math3d.js
@@ -2,6 +2,7 @@ import {
   glMatrix, mat2, mat2d, mat3, mat4, quat, quat2, vec2, vec3, vec4,
 } from '@libs/gl-matrix-min';
 import { LiteGraph } from '@/litegraph';
+import { LGraphNode } from '../core/LGraphNode';
 
 const global = typeof (window) !== 'undefined' ? window : typeof (self) !== 'undefined' ? self : globalThis;
