+++ tools/tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     /* Language and Environment */
     "target": "es2016",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
-    "module": "ES6",
+    "module": "ESNext",
+    "moduleResolution": "Bundler",
 
     "lib": ["ESNext", "DOM"],
